requirements.txt
neo4j
py-radix
pandas
plotly
IPython
